requirements.txt
langchain
langchain-community
pypdf
openai
gtts
edge-tts
groq
python-dotenv
nest_asyncio
ffmpeg 
moviepy
flask